+++ ZoneScouter/Core/UI/SectorInfoPanel.cs
@@ -23,8 +23,8 @@ public sealed class SectorInfoPanel {
   public ValueWithLabel SectorXY { get; private set; }
   public ValueWithLabel SectorZdoCount { get; private set; }
 
-  public ContentRow ZdoManagerContent { get; private set; }
-  public ValueWithLabel ZdoManagerNextId { get; private set; }
+  public ContentRow ZDOManagerContent { get; private set; }
+  public ValueWithLabel ZDOManagerNextId { get; private set; }
 
   public PanelDragger PanelDragger { get; private set; }
 
@@ -44,7 +44,7 @@ public sealed class SectorInfoPanel {
     PositionZ = new(PositionContent.Row.transform);
     PositionZ.Label.SetText("Z");
 
-    CreateCopyPositionButton();
+    CopyPositionButton = CreateCopyPositionButton(PositionContent.Row.transform);
 
     SectorContent = new(Panel.transform);
 
@@ -54,84 +54,76 @@ public sealed class SectorInfoPanel {
     SectorZdoCount = new(SectorContent.Row.transform);
     SectorZdoCount.Label.SetText("ZDOs");
 
-    ZdoManagerContent = new(Panel.transform);
+    ZDOManagerContent = new(Panel.transform);
 
-    ZdoManagerNextId = new(ZdoManagerContent.Row.transform);
-    ZdoManagerNextId.Label.SetText("NextId");
+    ZDOManagerNextId = new(ZDOManagerContent.Row.transform);
+    ZDOManagerNextId.Label.SetText("NextId");
 
     SetPanelStyle();
 
+    ToggleSectorContent(ShowSectorContent.Value);
     ToggleZDOManagerContent(ShowZDOManagerContent.Value);
 
     PanelDragger = Panel.AddComponent<PanelDragger>();
     PanelDragger.TargetRectTransform = Panel.GetComponent<RectTransform>();
   }
 
-  void CreateCopyPositionButton() {
-    CopyPositionButton = new(PositionContent.Row.transform);
-
-    CopyPositionButton.Cell.AddComponent<LayoutElement>()
-        .SetIgnoreLayout(true);
-
-    CopyPositionButton.Cell.GetComponent<RectTransform>()
-        .SetAnchorMin(new(1f, 0.5f))
-        .SetAnchorMax(new(1f, 0.5f))
-        .SetPivot(new(0f, 0.5f))
-        .SetPosition(new(10f, 0f))
-        .SetSizeDelta(new(80f, 0f));
-
-    CopyPositionButton.Label.text = "Copy";
-  }
-
-  public void ToggleCopyButtons(bool toggleOn) {
-    CopyPositionButton.Cell.gameObject.SetActive(toggleOn);
-  }
-
   public void SetPanelStyle() {
+    Background.SetColor(SectorInfoPanelBackgroundColor.Value);
+
     int fontSize = SectorInfoPanelFontSize.Value;
 
     PositionX.Label.SetFontSize(fontSize);
     PositionX.Value.SetFontSize(fontSize);
     PositionX.Value.SetColor(PositionValueXTextColor.Value);
     PositionX.FitValueToText("-00000");
-    PositionX.Background.SetColor(PositionValueXTextColor.Value.SetAlpha(0.1f));
+    PositionX.Background.SetColor(PositionValueXBackgroundColor.Value);
 
     PositionY.Label.SetFontSize(fontSize);
     PositionY.Value.SetFontSize(fontSize);
     PositionY.Value.SetColor(PositionValueYTextColor.Value);
     PositionY.FitValueToText("-00000");
-    PositionY.Background.SetColor(PositionValueYTextColor.Value.SetAlpha(0.1f));
+    PositionY.Background.SetColor(PositionValueYBackgroundColor.Value);
 
     PositionZ.Label.SetFontSize(fontSize);
     PositionZ.Value.SetFontSize(fontSize);
     PositionZ.Value.SetColor(PositionValueZTextColor.Value);
     PositionZ.FitValueToText("-00000");
-    PositionZ.Background.SetColor(PositionValueZTextColor.Value.SetAlpha(0.1f));
+    PositionZ.Background.SetColor(PositionValueZBackgroundColor.Value);
 
     SectorXY.Label.SetFontSize(fontSize);
     SectorXY.Value.SetFontSize(fontSize);
     SectorXY.Value.SetColor(PositionValueXTextColor.Value);
     SectorXY.FitValueToText("-123,-123");
-    SectorXY.Background.SetColor(PositionValueXTextColor.Value.SetAlpha(0.1f));
+    SectorXY.Background.SetColor(PositionValueXBackgroundColor.Value);
 
     SectorZdoCount.Label.SetFontSize(fontSize);
     SectorZdoCount.Value.SetFontSize(fontSize);
     SectorZdoCount.Value.SetColor(PositionValueYTextColor.Value);
     SectorZdoCount.FitValueToText("123456");
-    SectorZdoCount.Background.SetColor(PositionValueYTextColor.Value.SetAlpha(0.1f));
+    SectorZdoCount.Background.SetColor(PositionValueYBackgroundColor.Value);
 
-    ZdoManagerNextId.Label.SetFontSize(fontSize);
-    ZdoManagerNextId.Value.SetFontSize(fontSize);
-    ZdoManagerNextId.Value.SetColor(PositionValueZTextColor.Value);
-    ZdoManagerNextId.FitValueToText("1234567890");
-    ZdoManagerNextId.Background.SetColor(PositionValueZTextColor.Value.SetAlpha(0.1f));
+    ZDOManagerNextId.Label.SetFontSize(fontSize);
+    ZDOManagerNextId.Value.SetFontSize(fontSize);
+    ZDOManagerNextId.Value.SetColor(PositionValueZTextColor.Value);
+    ZDOManagerNextId.FitValueToText("1234567890");
+    ZDOManagerNextId.Background.SetColor(PositionValueZBackgroundColor.Value);
+  }
+
+  public void ToggleMenuComponents(bool toggleOn) {
+    PanelDragger.enabled = toggleOn;
+    CopyPositionButton.Cell.gameObject.SetActive(toggleOn);
+  }
+
+  public void ToggleSectorContent(bool toggleOn) {
+    SectorContent.Row.SetActive(toggleOn);
   }
 
   public void ToggleZDOManagerContent(bool toggleOn) {
-    ZdoManagerContent.Row?.SetActive(toggleOn);
+    ZDOManagerContent.Row.SetActive(toggleOn);
   }
 
-  GameObject CreatePanel(Transform parentTransform) {
+  static GameObject CreatePanel(Transform parentTransform) {
     GameObject panel = new("SectorInfo.Panel", typeof(RectTransform));
     panel.transform.SetParent(parentTransform, worldPositionStays: false);
 
@@ -147,13 +139,31 @@ public sealed class SectorInfoPanel {
 
     panel.AddComponent<Image>()
         .SetType(Image.Type.Sliced)
-        .SetSprite(UIBuilder.CreateSuperellipse(200, 200, 12))
+        .SetSprite(UISpriteBuilder.CreateSuperellipse(200, 200, 12))
         .SetColor(SectorInfoPanelBackgroundColor.Value);
 
     panel.AddComponent<CanvasGroup>()
         .SetBlocksRaycasts(true);
 
     return panel;
+  }
+
+  static ButtonCell CreateCopyPositionButton(Transform parentTransform) {
+    ButtonCell buttonCell = new(parentTransform);
+
+    buttonCell.Cell.AddComponent<LayoutElement>()
+        .SetIgnoreLayout(true);
+
+    buttonCell.Cell.GetComponent<RectTransform>()
+        .SetAnchorMin(new(1f, 0.5f))
+        .SetAnchorMax(new(1f, 0.5f))
+        .SetPivot(new(0f, 0.5f))
+        .SetSizeDelta(new(80f, 0f))
+        .SetPosition(new(10f, 0f));
+
+    buttonCell.Label.text = "Copy";
+
+    return buttonCell;
   }
 
   public sealed class ContentRow {
@@ -163,7 +173,7 @@ public sealed class SectorInfoPanel {
       Row = CreateChildRow(parentTransform);
     }
 
-    GameObject CreateChildRow(Transform parentTransform) {
+    static GameObject CreateChildRow(Transform parentTransform) {
       GameObject row = new("Row", typeof(RectTransform));
       row.transform.SetParent(parentTransform, worldPositionStays: false);